// estore-ui/src/app/interfaces/checkout-data.ts
export interface CheckoutData {
  userID: number;
  firstName: string;
  lastName: string;
  address: string;
  city: string;
  state: string;
  country: string;
  zipCode: number;
  email: string;
  phoneNumber: string;
  creditCardNumber: string;
  creditCardExpiration: string;
  creditCardCVC: number;
  creditCardHolder: string;
  creditCardZipCode: number;
}